extract new provider for `sourceDeclaration`
