+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotype/snippet/forkosourcedeclarationprovider/not-nullable-import-alias-type.kttxt
@@ -1,0 +1,3 @@
+import com.lemonappdev.konsist.testdata.SampleObject as ImportAlias
+
+class SampleClass1(val sampleProperty1: ImportAlias)
