+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotype/snippet/forkosourcedeclarationprovider/not-nullable-typealias-type.kttxt
@@ -1,0 +1,5 @@
+import com.lemonappdev.konsist.testdata.SampleObject
+
+class SampleClass1(val sampleProperty1: SampleTypeAlias)
+
+typealias SampleTypeAlias = (SampleObject) -> Unit
