+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotype/snippet/forkosourcedeclarationprovider/nullable-class-type.kttxt
@@ -1,0 +1,3 @@
+import com.lemonappdev.konsist.testdata.SampleType
+
+class SampleClass1(val sampleProperty1: SampleType?)
