+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/type/kotype/snippet/forkosourcedeclarationprovider/nullable-kotlin-collection-type.kttxt
@@ -1,0 +1,1 @@
+class SampleClass1(val sampleProperty1: List<String>?)
